[Grid] Move GridSpan to its own file. https://bugs.webkit.org/show_bug.cgi?id=293491 rdar://151927977
Reviewed by Vitor Roriz.

The header file it currently lives in, GridArea.h, is a bit bloated. We
can move this to its own header and source file which helps with
organization and build speed.

Canonical link: https://commits.webkit.org/295462@main
