+++ Source/WebCore/svg/properties/SVGAnimationAdditiveValueFunctionImpl.h
@@ -29,6 +29,7 @@
 #include "FloatRect.h"
 #include "SVGAngleValue.h"
 #include "SVGAnimationAdditiveValueFunction.h"
+#include "SVGElement.h"
 #include "SVGLengthContext.h"
 #include "SVGLengthValue.h"
 #include "SVGPathByteStream.h"
